// src/monument/types.ts
export interface MonumentStructure {
  id: string;
  name: string;
  description: string;
  imageUrl: string;
  country: string;
  city: string;
}

export type MonumentCommonData = {
  name: string;
  description: string;
  imageUrl: string;
  country: string;
  city: string;
};

export type Location = {
  country: string;
  city: string;
};